Replace default formatting example with gaussian plot
